requirements.txt
langchain==0.3.15
langchain_core==0.3.31
langchain_openai==0.3.2
langchain-anthropic
langchain-groq
lxml==5.1.0
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
langchain_community
faiss-cpu
spacy
nltk
# Added for DeepSeek-V3 models support
transformers>=4.51.3
torch>=2.2.2
langchain-huggingface>=0.1.2
sentence-transformers>=4.1.0
pandas
json5
demjson3